Add password regExp, change folder structure,  refactor
