+++ src/controllers/userSignUpHandler.js
@@ -1,0 +1,5 @@
+const handler = (request, reply) => {
+  reply('User Signed Up!').code(201);
+};
+
+module.exports = handler;
